Integrating game setup with main menu
Networking components required to host/join games now
integrated with the main menu. On successful creation
of a full game, each client has knowledge of the other
clients (ipv4, port, is host, pen color). At this point,
integration with the in-game components/ui is ready to
begin.

NOTE: to run the main application, enter: mvn compile exec:java
in the command line with the root project directory as your cwd